Simon says done + fix
Flower now plays exemple melody when player go near it.
Charge and plant count is displayed in the top left corner of the screen.
Player can change a charge into a plant by playing the notes in the same order than the flower.

Fix material assignement for sound particle system.