Multiple melodies recording + Visual changes
Melodies
5 melodies are tracked at all time when the player plays a note.
Each melody can be heard frome different plants.
Each plant is a flower of a color that display the difficulty of the melody that it plays.
It goes from easiest to hardest in this order :  light blue, light green, yellow, orange, red.

Visual
Added the 5 different types of charges and plants in the UI.
Added model for the flowers.
Added rocks model and texture.
Added the 5 types of flowers in the scene in an increasing order of difficulty.
Added rocks in the scene to fill the empty plain.
Tweaked the sound particle system especially for the flowers.